requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2.3
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-writable-nested==0.7.2
drf-yasg==1.21.10
fonttools==4.58.4
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
swagger-spec-validator==3.0.4
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
uritemplate==4.2.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
